requirements.txt
absl-py==2.2.2
aiosignal==1.3.1
ale-py==0.10.2
annotated-types==0.7.0
astunparse==1.6.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.0.4
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.71.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.1.1
h5py==3.13.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
jax-jumpy==1.0.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.9
networkx==3.2.1
numpy==1.23.5
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==24.1
pandas==2.2.2
pettingzoo==1.24.3
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.20.3
psutil==6.0.0
pyarrow==16.1.0
pydantic==1.10.12
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.0.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.13.1
sentry-sdk==2.7.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
stable_baselines3==2.6.0
SuperSuit==3.10.0
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tifffile==2024.6.18
tinyscaler==1.2.8
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.67.1
typer==0.12.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wandb==0.19.10
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.19.2
